fix(rest): use new path service validation in both v2/v4 in order to check either virtualhost or path
