+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/rest/api/service/v4/validation/PathValidationService.java
@@ -15,7 +15,6 @@
  */
 package io.gravitee.rest.api.service.v4.validation;
 
-import io.gravitee.definition.model.v4.listener.Listener;
 import io.gravitee.definition.model.v4.listener.http.Path;
 import io.gravitee.rest.api.service.common.ExecutionContext;
 import java.util.List;
@@ -24,6 +23,8 @@ import java.util.List;
  * @author Guillaume LAMIRAND (guillaume.lamirand at graviteesource.com)
  * @author GraviteeSource Team
  */
-public interface ListenerValidationService {
-    List<Listener> validateAndSanitize(ExecutionContext executionContext, String apiId, List<Listener> listeners);
+public interface PathValidationService {
+    List<Path> validateAndSanitizePaths(ExecutionContext executionContext, String apiId, List<Path> paths);
+
+    String sanitizePath(String path);
 }
